+++ VadereState/src/org/vadere/state/attributes/processor/AttributesRelevantPedestriansProcessor.java
@@ -1,4 +1,0 @@
-package org.vadere.state.attributes.processor;
-
-public class AttributesRelevantPedestriansProcessor extends AttributesProcessor {
-}